Code refactoring: separate schema classification.
